+++ react/src/js/components/Consonant/CardsCarousel/__tests__/CardsCarousel.spec.js
@@ -318,4 +318,89 @@ describe('CardsCarousel comprehensive behaviors', () => {
     const prevBtn = c.querySelector('[name="previous"]');
     expect(document.activeElement).toBe(prevBtn);
   });
+
+  // Test for carouselType
+  test('should set carouselType to Modern when animationStyle is incrementalModern', async () => {   
+    const cfg = JSON.parse(JSON.stringify(config));
+    cfg.collection.layout.container = 'carousel';
+    cfg.pagination.animationStyle = 'incrementalModern';
+    await act(async () => {
+      ({ container } = render(<Container config={cfg} />));
+    });
+    const carousel = container.querySelector('.consonant-Wrapper-collection');
+    expect(carousel.classList.contains('modern-carousel')).toBe(true);
+  });
+
+  test('should set carouselType to Modern when animationStyle is pagedModern', async () => {   
+    const cfg = JSON.parse(JSON.stringify(config));
+    cfg.collection.layout.container = 'carousel';
+    cfg.pagination.animationStyle = 'pagedModern';
+    await act(async () => {
+      ({ container } = render(<Container config={cfg} />));
+    });
+    const carousel = container.querySelector('.consonant-Wrapper-collection');
+    expect(carousel.classList.contains('modern-carousel')).toBe(true);
+  });
+
+  test('should set carouselType to default when animationStyle is incremental', async () => {   
+    const cfg = JSON.parse(JSON.stringify(config));
+    cfg.collection.layout.container = 'carousel';
+    cfg.pagination.animationStyle = 'incremental';
+    await act(async () => {
+      ({ container } = render(<Container config={cfg} />));
+    });
+    const carousel = container.querySelector('.consonant-Wrapper-collection');
+    expect(carousel.classList.contains('modern-carousel')).toBe(false);
+  });
+
+  test('should set carouselType to default when animationStyle is paged', async () => {   
+    const cfg = JSON.parse(JSON.stringify(config));
+    cfg.collection.layout.container = 'carousel';
+    cfg.pagination.animationStyle = 'paged';
+    await act(async () => {
+      ({ container } = render(<Container config={cfg} />));
+    });
+    const carousel = container.querySelector('.consonant-Wrapper-collection');
+    expect(carousel.classList.contains('modern-carousel')).toBe(false);
+  });
+
+  test('should NOT set carousel class to --light', async () => {
+    const cfg = JSON.parse(JSON.stringify(config));
+    cfg.collection.layout.container = 'carousel';
+    cfg.pagination.animationStyle = 'pagedModern';
+    await act(async () => {
+      ({ container } = render(<Container config={cfg} />));
+    });
+    const carousel = container.querySelector('.consonant-Wrapper-collection');
+    expect(carousel.classList.contains('modern-carousel--light')).toBe(false);
+  });
+
+  test('should set carousel class to --light', async () => {
+    const cfg = JSON.parse(JSON.stringify(config));
+    cfg.collection.layout.container = 'carousel';
+    cfg.pagination.animationStyle = 'pagedModern-light';
+    await act(async () => {
+      ({ container } = render(<Container config={cfg} />));
+    });
+    const carousel = container.querySelector('.consonant-Wrapper-collection');
+    expect(carousel.classList.contains('modern-carousel--light')).toBe(true);
+  });
+
+  test('should set modern carousel with incremental pagination', async () => {
+    const cfg = JSON.parse(JSON.stringify(config));
+    cfg.pagination.animationStyle = 'incrementalModern-light';
+    cfg.collection.layout = { container: 'carousel', gutter: '3x', type: '2up' };
+
+    await act(async () => {
+      ({ container } = render(<Container config={cfg} />));
+    });
+    
+    const carousel = container.querySelector('.consonant-Container--carousel');
+    const initialScroll = carousel.scrollLeft;
+    fireEvent.click(container.querySelector('[name="next"]'));
+    expect(carousel.scrollLeft).toBe(initialScroll + 595); // cardWidth (579) + gridGap (16)
+
+    fireEvent.click(container.querySelector('[name="previous"]'));
+    expect(carousel.scrollLeft).toBe(initialScroll);
+  });
 });